deployment/tts-api/requirements.txt
absl-py==2.3.1
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
argbind==0.3.9
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.3.0
audioread==3.0.1
babel==2.16.0
bibtexparser==2.0.0b8
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
clldutils==3.24.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.3
contourpy==1.3.1
csvw==3.5.1
cycler==0.12.1
datasets==3.2.0
debugpy==1.8.16
decorator==5.1.1
descript-audio-codec==1.0.0
descript-audiotools==0.7.2
dill==0.3.8
dlinfo==1.2.1
docstring_parser==0.17.0
dotwiz==0.4.0
einops==0.8.1
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.6
ffmpy==0.6.1
filelock==3.16.1
fire==0.7.0
flatbuffers==24.12.23
flatten-dict==0.4.2
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.9.0
future==1.0.0
g2p-id @ git+https://github.com/Wikidepia/g2p-id@309063b4c06be2e67482e3c1c886b1973af86842
gradio==5.41.1
gradio_client==1.11.0
groovy==0.1.2
grpcio==1.74.0
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.10
imageio==2.36.1
importlib_resources==6.5.2
ipykernel==6.30.1
ipython==8.37.0
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
julius==0.2.7
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
language-tags==1.2.0
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
loguru==0.7.3
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.4
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
nari-tts @ git+https://github.com/nari-labs/dia.git@92206cfb33fbb599166c134d20d247b3c6bd5919
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numpy==2.2.6
onnxruntime==1.20.1
orjson==3.11.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
phonemizer==3.3.0
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.51
propcache==0.2.1
protobuf==3.19.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyheck==0.1.5
pylatexenc==2.10
pyloudnorm==0.1.1
pyparsing==3.2.1
pystoi==0.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
pyzmq==27.0.1
randomname==0.2.1
rdflib==7.1.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rich==14.1.0
rpds-py==0.22.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
ruff==0.12.7
sacremoses==0.1.1
safehttpx==0.1.6
safetensors==0.6.1
scikit-learn==1.6.0
scipy==1.15.0
segments==2.2.1
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
speechbrain==1.0.2
sse-starlette==2.2.1
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
threadpoolctl==3.5.0
tokenizers==0.21.0
tomlkit==0.13.3
torch==2.6.0
torch-stoi==0.2.3
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.18.3